:sparkles:Feat: update functions related to sellerpage&review #5 #16 #19
 -update productdetail.html & javascript
 -add writereview.html & javascript
