+++ header.html
@@ -25,8 +25,8 @@
             <img id="cart" style='cursor:pointer' src="static/images/shopping-cart.png" alt="cart"
                 onclick="location.href='cart.html'">
             <ul id="intro" style='cursor:pointer' onclick="location.href='/user_detail_page.html';"> </ul>
-            <ul id="charge" style='cursor:pointer' onclick="location.href='pointcharge.html';"> 포인트충전</ul>
-            <ul id="mypage" style='cursor:pointer' onclick="location.href='mypage.html';"> 마이페이지 </ul>
+            <!-- <ul id="charge" style='cursor:pointer' onclick="location.href='pointcharge.html';"> 포인트충전</ul> -->
+            <!-- <ul id="mypage" style='cursor:pointer' onclick="location.href='mypage.html';"> 마이페이지 </ul> -->
             <ul id="sellerpage" style='cursor:pointer' onclick="location.href='seller.html';"> 판매자페이지 </ul>
             <ul id="logout" style='cursor:pointer'> 로그아웃 &nbsp; &nbsp; </a>
         </div>
