+++ static/css/header.css
@@ -18,7 +18,8 @@ body {
 }
 
 a {
-    color: #333;
+    /* color: #333; */
+    color:#522F18;
     text-decoration: none;
 }
 
@@ -87,6 +88,7 @@ header {
 
 .top-category {
     display: flex;
+    color:#522F18;
 }
 
 nav {
@@ -100,7 +102,7 @@ nav a {
     display: block;
     width: 20%;
     padding: .75em 0;
-    color: #333;
+    color: #522F18;
     text-decoration: none;
     text-align: center;
 }
@@ -111,7 +113,7 @@ nav a {
     bottom: -2px;
     width: 20%;
     height: 2px;
-    background: #333;
+    background: #522F18;
     transition: all .3s ease-in-out;
 }
 
@@ -168,7 +170,7 @@ nav a:nth-child(5):hover~.nav-underline {
     height: 32px;
     text-align: center;
     line-height: 30px;
-    background: #f5f5f5;
+    background: white;
     color: #444;
     cursor: pointer;
     border: 0;
