+++ static/css/productdetail.css
@@ -1,19 +1,63 @@
+@font-face {
+    font-family: 'GangwonEdu_OTFBoldA';
+    src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2201-2@1.0/GangwonEdu_OTFBoldA.woff') format('woff');
+    font-weight: normal;
+    font-style: normal;
+}
+
 .section3 {
+    font-family: 'GangwonEdu_OTFBoldA';
     background-color: #C4A492;
     height: 100%;
-    
     width: 100%;
     padding-bottom: 100px;
 }
 
-#post-card {
-    width:660px;
-    /* display: flex;
-    justify-content: center; */
+#review-List {
+    width:500px;
+    margin-left:400px;
 }
-#profile {
-    /* display: flex;
-    justify-content: center; */
-    width: 600px;
+.card {
+    margin-bottom: 10px;
+}
+#post-card {
+    padding-top: 30px;
+}
+#imagePut {
+    margin: 50px 0 0 210px;
+    width : 70%;
+    height: 50%;
 }
 
+#product-title{
+    padding-top: 30px;
+    margin: 40px 0 30px 210px;
+}
+
+#product-content {
+    margin: 50px 0 40px 210px;
+    font-size: 30px;
+}
+#product-price{
+    margin: 50px 0 40px 210px;
+    font-size: 30px;
+}
+#product-amount{
+    margin: 50px 0 40px 210px;
+    font-size: 30px;
+}
+#top-image {
+    margin-left: 180px;
+}
+
+#edit-btn {
+    --bs-btn-padding-y: 0.7rem; 
+    --bs-btn-padding-x: 1.7rem; 
+    --bs-btn-font-size: 1.5rem;
+    margin-left: 760px;
+}
+#delete-btn {
+    --bs-btn-padding-y: 0.7rem; 
+    --bs-btn-padding-x: 1.7rem; 
+    --bs-btn-font-size: 1.5rem;
+}
