+++ static/js/bill.js
@@ -1,5 +1,6 @@
 import { BACK_BASE_URL, FRONT_BASE_URL, getSubscribeView, getUserProfileAPIView, getBillList, billToCart } from './api.js'
 
+
 async function renderBillList() {
     const bills = await getBillList();
     const billBox = document.getElementById('billLists');
