+++ static/js/productregistration.js
@@ -133,32 +133,6 @@ export async function setEventListener() {
       
 }
 
-// 판매자정보
-async function sellerProfile() {
-    const payload = localStorage.getItem("payload");
-    const payload_parse = JSON.parse(payload);
-    const user_id = payload_parse.user_id //로그인한 유저id
-
-    const seller_data = await getSellerPermissionAPIView(user_id)
-    console.log("판매자정보", seller_data)
-
-    let today = new Date();
-    let year = today.getFullYear();
-    let month = String(today.getMonth() + 1).padStart(2, '0'); //두자리되도록 앞에0채우기
-    let date = String(today.getDate()).padStart(2, '0'); //두자리되도록 앞에0채우기
-    today = `${year}-${month}-${date}`;
-    // console.log("today", today)
-
-    
-
-
-    if (seller_data['company_img']) { //로고 이미지가 존재한다면
-        document.getElementById("company-img").setAttribute("src", `${BACK_BASE_URL}` + seller_data['company_img'])
-    }
-    document.getElementById("company-name").innerText = seller_data["company_name"]
-    document.getElementById("owner-name").innerText = seller_data["business_owner_name"]
-}
-sellerProfile()
 
 window.onload = async function() {
     const urlParams = new URLSearchParams(window.location.search);
