+++ writereview.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>리뷰 작성</title>
+    <!-- Bootstrap -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
+    <!-- CSS -->
+    <link rel="stylesheet" href="static/css/header.css">
+    <link rel="stylesheet" href="static/css/footer.css">
+    <link rel="stylesheet" href="static/css/writereview.css">
+</head>
+<header></header>
+<body>
+    <div class = "section2">
+        
+        <form>
+            <img src="/static/images/리뷰.gif" id="top-image">
+            <fieldset class="postField container" name="postField" id="postField2">
+                <div class="title-box">
+                    <input type="text" class="form-control" id="review-title" placeholder="제목">
+                </div>
+                <select class="form-select" id="give-star" aria-label="Default select example">
+                    <option selected>상품 별점 주기</option>
+                    <option value="1">⭐</option>
+                    <option value="2">⭐⭐</option>
+                    <option value="3">⭐⭐⭐</option>
+                    <option value="4">⭐⭐⭐⭐</option>
+                    <option value="5">⭐⭐⭐⭐⭐</option>
+                </select>
+                <div class="mb-3">
+                    <input class="form-control" type="file" id="formFile">
+                </div>
+                <div class="input-group" id="reviewBox">
+                    <span class="input-group-text">내용</span>
+                    <textarea class="form-control" aria-label="With textarea" id="review-content"></textarea>
+                </div>
+                <button type="button" class="btn btn-secondary" id="reviewsubmitbutton"
+                        style="--bs-btn-padding-y: 0.7rem; --bs-btn-padding-x: 1.7rem; --bs-btn-font-size: 1.5rem;">작성하기</button>
+                <button type="button" class="btn btn-secondary" id="update-submit"
+                style="--bs-btn-padding-y: 0.7rem; --bs-btn-padding-x: 1.7rem; --bs-btn-font-size: 1.5rem;">수정하기</button>
+            </fieldset>
+        </form>
+    </div>
+</body>
+<footer></footer>
+<!-- script -->
+<script src="/static/js/api.js" type="module"></script>
+<script src="/static/js/loader.js" type="module"></script>
+<script src="/static/js/writereview.js" type="module"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
+    integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
+    crossorigin="anonymous"></script>
+</html>
